% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarySearch.R
\name{BinarySearch}
\alias{BinarySearch}
\title{This function calculates the lambda for desired sum of absolute value
for the loadings}
\usage{
BinarySearch(argu, sumabs)
}
\arguments{
\item{argu}{input vector}

\item{sumabs}{value of sum of absolute values for loadings desired}
}
\description{
This function calculates the lambda for desired sum of absolute value
for the loadings
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
