% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.decomp.R
\name{dual.decomp}
\alias{dual.decomp}
\title{This function converts the principal component decomposition to the dual
supervised form}
\usage{
dual.decomp(X, L, npc, n, p)
}
\arguments{
\item{X}{nxp matrix of predictors}

\item{L}{nxn response kernel}

\item{npc}{number of desired principal components}

\item{n}{an integer storing the number response observations}

\item{p}{an integer storing the number of predictors}
}
\description{
This function converts the principal component decomposition to the dual
supervised form
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
